requirements.txt
psutil>=5.9.0
Pillow>=9.0.0
pyserial>=3.5
rumps>=0.4.0
Flask>=2.0.0
matplotlib>=3.5.0
bcrypt>=4.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
